requirements.txt
GitPython
numpy
pillow